Module dprint_plugin_typescript::configuration

Structs

Configuration

ConfigurationBuilder

TypeScript formatting configuration builder.

Enums

BracePosition

Where to place the opening brace.

MemberSpacing

How to space members.

NextControlFlowPosition

Where to place the next control flow within a control flow statement.

OperatorPosition

Where to place the operator for expressions that span multiple lines.

QuoteStyle

How to decide to use single or double quotes.

SemiColonOrComma

Whether to use semi-colons or commas.

SemiColons

Semi colon possibilities.

SingleBodyPosition

Where to place the expression of a statement that could possibly be on one line (ex. if (true) console.log(5);).

SortOrder

The kind of sort ordering to use.

TrailingCommas

Trailing comma possibilities.

UseBraces

If braces should be used or not in certain scenarios.

UseParentheses

Whether to use parentheses around a single parameter in an arrow function.

Functions

resolve_config

Resolves configuration from a collection of key value strings.